pgdbencryption/requirements.txt
docker-compose
flask
flask-shell-ipython
flask-sqlalchemy
flask-wtf
psycopg2-binary
pytest-flask
SQLAlchemy-utils